+++ packages/core/src/extensions/commands.ts
@@ -12,6 +12,7 @@ import * as extendMarkRange from '../commands/extendMarkRange'
 import * as first from '../commands/first'
 import * as focus from '../commands/focus'
 import * as insertHTML from '../commands/insertHTML'
+import * as insertNode from '../commands/insertNode'
 import * as insertText from '../commands/insertText'
 import * as joinBackward from '../commands/joinBackward'
 import * as joinForward from '../commands/joinForward'
@@ -58,6 +59,7 @@ export { extendMarkRange }
 export { first }
 export { focus }
 export { insertHTML }
+export { insertNode }
 export { insertText }
 export { joinBackward }
 export { joinForward }
@@ -109,6 +111,7 @@ export const Commands = Extension.create({
       ...first,
       ...focus,
       ...insertHTML,
+      ...insertNode,
       ...insertText,
       ...joinBackward,
       ...joinForward,